iot_ids_llm_rag/LLM/requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
tensorflow
torch
transformers
sentence-transformers
faiss-cpu
datasets
accelerate
imblearn
